Convert double to BigDecimal in Functions and EvaluateString

+++ src/Functions/Function.java
@@ -1,5 +1,7 @@
 package Functions;
 
+import java.math.BigDecimal;
+
 /**
  * The Function interface represents a mathematical function.
  * It provides methods to evaluate the value, differentiate, integrate,
@@ -13,7 +15,7 @@ public interface Function {
      * @param x The x-coordinate at which to evaluate the function.
      * @return The value of the function at the given x-coordinate.
      */
-    double getValueAt(double x);
+    BigDecimal getValueAt(BigDecimal x);
 
     /**
      * Computes the value of the derivative of the function at the given x-coordinate.
@@ -22,7 +24,7 @@ public interface Function {
      * @param rank The order of the derivative (e.g., 1 for first derivative, 2 for second derivative, and so on).
      * @return The value of the derivative of the function at the given x-coordinate.
      */
-    double getDiffAt(double x, int rank);
+    BigDecimal getDiffAt(BigDecimal x, int rank);
 
     /**
      * Computes the value of the integral of the function at the given x-coordinate.
@@ -31,7 +33,7 @@ public interface Function {
      * @param rank The order of the integral (e.g., 1 for first integral, 2 for second integral, and so on).
      * @return The value of the integral of the function at the given x-coordinate.
      */
-    double getIntegralAt(double x, int rank);
+    BigDecimal getIntegralAt(BigDecimal x, int rank);
 
     /**
      * Returns the string representation of the function.
